build: consolidate all deps under vcpkg with custom triplets
- Add build/vcpkg-triplets/ with x64-mingw-cross, x86-mingw-cross, x64-linux-keeperfx
- Update CMakePresets.json: linux-base and cross-base now use vcpkg toolchain;
  windows-x64/x86 presets use mingw-cross triplets (no direct toolchain override);
  VCPKG_OVERLAY_TRIPLETS set globally in base preset
- Rewrite deps/CMakeLists.txt: remove all FetchContent/prebuilt archive/pkg-config
  branching; pure find_package() for all deps via vcpkg
- Update vcpkg.json: platform guards for Vulkan packages (windows-only);
  move curl before Vulkan entries
- Simplify Dependencies.cmake: unified curl detection (vcpkg for all platforms);
  remove MinGW-specific sysroot search paths
- Update docker/mingw64: strip SDL2 and curl prebuilt install blocks; add zip/python3/nasm
- Update docker/mingw32: strip SDL2 prebuilts and kfx-deps cache; add zip/python3/nasm
- Update docker/linux: replace installed dev packages with vcpkg build prerequisites
  (X11, Wayland, ALSA, codec headers that vcpkg ports need to compile); add nasm/python3/zip
- Update ci-build.yml: add vcpkg bootstrap step + binary cache (x-gha) to Docker
  matrix jobs; remove manual apt install steps for SDL2/curl

Source fixes (all warning/error fixes from this session):
- net_lan.c: format-truncation %s -> %.*s for session name and join address
- lvl_script_commands.c: 6x format-truncation fixes for snprintf calls
- lua_base.c: buffer sizes enlarged (PATH_LENGTH*4, DISKPATH_SIZE+32)
- main.cpp: param_buffer truncation fix
- bflib_fmvids.cpp: -(long)sizeof to fix unsigned negation overflow
- globals.h: PRIu64 -> %llu with (unsigned long long) cast for MinGW
- TileAtlasPacker.cpp, LensManager.cpp: %zu/%lu -> %u/%llu casts
- kfx/lense: RENDERER_OPENGL_ENABLED guards on all 4 effect headers+cpps;
  else SYNCDBG -> else { SYNCDBG } to fix empty-body warnings
- platform/PlatformWindows.cpp: sstream>> typo; FileWrite return value
- platform/WindowSystemSDL.cpp: remove extern C from lbWindow definition
- platform/PlatformManager.cpp: Tracy header and calls guarded by TRACY_ENABLE
- renderer/RendererManager.cpp: s_rebuild_deferred/s_spriteAtlas inside GL guard
- abs() -> labs() for MapCoordDelta and map coordinate fields (17 files)

+++ src/bflib_fmvids.cpp
@@ -1408,7 +1408,7 @@ short anim_open(char *fname, int arg1, short arg2, int width, int height, int bp
 				return false;
 			}
 		} else {
-			LbFileSeek(animation.inpfhndl, -sizeof(AnimFLIChunk), Lb_FILE_SEEK_CURRENT);
+			LbFileSeek(animation.inpfhndl, -(long)sizeof(AnimFLIChunk), Lb_FILE_SEEK_CURRENT);
 		}
 		animation.frame_count = 0;
 	}
