+++ src/kfx/lense/LensManager.cpp
@@ -463,7 +463,7 @@ TbBool LensManager::AllocateBuffers()
     m_spare_screen_memory = (unsigned char*)KfxCalloc(buffer_size, sizeof(unsigned char));
     
     if (m_lens_memory == nullptr || m_spare_screen_memory == nullptr) {
-        ERRORLOG("Failed to allocate lens buffers (%lu bytes)", buffer_size * sizeof(uint32_t));
+        ERRORLOG("Failed to allocate lens buffers (%llu bytes)", (unsigned long long)(buffer_size * sizeof(uint32_t)));
         FreeBuffers();
         return false;
     }
